:recycle: 将单个哔哩哔哩文件拆开
