+++ nonebot_bison/platform/bilibili/__init__.py
@@ -1,0 +1,3 @@
+from .platforms import Bilibili as Bilibili
+from .platforms import Bilibililive as Bilibililive
+from .platforms import BilibiliBangumi as BilibiliBangumi
